AI-Human-Attribute-Analyzer/requirements.txt
streamlit>=1.31.0
google-generativeai>=0.3.0
Pillow>=10.0.0
python-dotenv>=1.0.0
protobuf
google-auth
pillow  